Add PETSCII to Unicode and Unicode to PETSCII support
This is a major refactoring with basic PETSCII to Unicode and Unicode
to PETSCII working.

There are still some changes that need to be done with configuration
and ergonomics of easy PetsciiString creation.  But this gets this
support out there now.

There is also duplicate data in the config and the config needs to be
refactored.